done changes that are mentioned
i had made all the changes that are advised to me. regarding encircling the date that is marked we have to use third party lib which i mentioned in my comment. I even tried using that library but on doing so the app is not behaving well. android devices cannot percepts that library and showing errors.